# openapi.yaml
components:
  schemas:
...
    EmailCheckRequest:
      description: Email availability check request
      properties:
        email:
          description: Email to check availability
          title: Email
          type: string
      required:
      - email
      title: EmailCheckRequest
      type: object
    EmailCheckResponse:
      description: Email availability check response
      properties:
        available:
          description: Whether email is available for registration
          title: Available
          type: boolean
        email:
          description: Email that was checked
          title: Email
          type: string
        message:
          description: Human readable message about availability
          title: Message
          type: string
      required:
      - email
      - available
      - message
      title: EmailCheckResponse
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    RefreshTokenRequest:
      description: Refresh token request schema
      properties:
        refresh_token:
          description: JWT refresh token
          title: Refresh Token
          type: string
      required:
      - refresh_token
      title: RefreshTokenRequest
      type: object
...
    TokenResponse:
      description: JWT token response schema
      properties:
        access_token:
          description: JWT access token
          title: Access Token
          type: string
        expires_in:
          description: Token expiration in seconds
          title: Expires In
          type: integer
        refresh_token:
          description: JWT refresh token
          title: Refresh Token
          type: string
        token_type:
          default: bearer
          description: Token type
          title: Token Type
          type: string
        user:
          $ref: '#/components/schemas/UserResponse'
          description: User information
      required:
      - access_token
      - refresh_token
      - expires_in
      - user
      title: TokenResponse
      type: object
...
    UserCreate:
      description: Simple user creation schema for auth
      properties:
        email:
          description: User email address
          title: Email
          type: string
        full_name:
          description: User's full name
          title: Full Name
          type: string
        organization_name:
          anyOf:
          - maxLength: 255
            type: string
          - type: 'null'
          description: Organization/company name
          title: Organization Name
        password:
          description: User password
          minLength: 8
          title: Password
          type: string
      required:
      - email
      - full_name
      - password
      title: UserCreate
      type: object
    UserLogin:
      description: Simple user login schema
      properties:
        email:
          description: User email
          title: Email
          type: string
        password:
          description: User password
          title: Password
          type: string
      required:
      - email
      - password
      title: UserLogin
      type: object
    UserMeResponse:
      description: Current user profile information for /me endpoint
      properties:
        avatar_url:
          anyOf:
          - type: string
          - type: 'null'
          description: Avatar URL
          title: Avatar Url
        created_at:
          description: Creation timestamp
          format: date-time
          title: Created At
          type: string
        email:
          description: User email
          title: Email
          type: string
        full_name:
          anyOf:
          - type: string
          - type: 'null'
          description: Full name
          title: Full Name
        id:
          description: Unique identifier
          format: uuid
          title: Id
          type: string
        is_active:
          description: Whether user is active
          title: Is Active
          type: boolean
        is_verified:
          description: Whether email is verified
          title: Is Verified
          type: boolean
        language:
          description: User language
          title: Language
          type: string
        last_login_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          description: Last login timestamp
          title: Last Login At
        organization_domain:
          anyOf:
          - type: string
          - type: 'null'
          description: Organization domain
          title: Organization Domain
        organization_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          description: Organization ID
          title: Organization Id
        organization_name:
          anyOf:
          - type: string
          - type: 'null'
          description: Organization name
          title: Organization Name
        organization_plan:
          anyOf:
          - type: string
          - type: 'null'
          description: Organization plan
          title: Organization Plan
        organization_timezone:
          anyOf:
          - type: string
          - type: 'null'
          description: Organization timezone
          title: Organization Timezone
        preferences:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          description: User preferences
          title: Preferences
        role:
          $ref: '#/components/schemas/UserRoleSchema'
          description: User role
        timezone:
          description: User timezone
          title: Timezone
          type: string
        updated_at:
          description: Last update timestamp
          format: date-time
          title: Updated At
          type: string
      required:
      - created_at
      - updated_at
      - id
      - email
      - role
      - is_active
      - timezone
      - language
      - is_verified
      title: UserMeResponse
      type: object
    UserResponse:
      description: Simple user response schema
      properties:
        created_at:
          description: Account creation timestamp
          format: date-time
          title: Created At
          type: string
        email:
          description: User email
          title: Email
          type: string
        full_name:
          anyOf:
          - type: string
          - type: 'null'
          description: User's full name
          title: Full Name
        id:
          description: User ID
          format: uuid
          title: Id
          type: string
        is_active:
          description: Whether user is active
          title: Is Active
          type: boolean
        last_login_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          description: Last login timestamp
          title: Last Login At
      required:
      - id
      - email
      - full_name
      - is_active
      - created_at
      title: UserResponse
      type: object
    UserRoleSchema:
      description: User roles enum schema
      enum:
      - admin
      - manager
      - agent
      - user
      - api_user
      title: UserRoleSchema
      type: string
    UserUpdate:
      description: Simple user update schema
      properties:
        full_name:
          anyOf:
          - type: string
          - type: 'null'
          description: User's full name
          title: Full Name
        password:
          anyOf:
          - minLength: 8
            type: string
          - type: 'null'
          description: New password
          title: Password
      title: UserUpdate
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    HTTPBearer:
      scheme: bearer
      type: http
info:
  description: "\n        # AI Ticket Creator Backend API\n        \n        Comprehensive\
    \ API for AI-powered ticket creation and management with Chrome extension support.\n\
    \        \n        ## Features\n        \n        - **AI-Powered Ticket Creation**:\
    \ Smart ticket creation with context analysis\n        - **File Processing**:\
    \ Upload and process multiple file types with AI transcription/OCR\n        -\
    \ **Real-time Notifications**: WebSocket-based real-time updates\n        - **Third-party\
    \ Integrations**: Salesforce, Jira, ServiceNow, Zendesk, GitHub, Slack, Teams,\
    \ Zoom\n        - **Chrome Extension Support**: Optimized for browser extension\
    \ workflows\n        - **Multi-Tenant Organization Support**: Complete organization\
    \ isolation for enterprise deployments\n        \n        ## Quick Start Guide\n\
    \        \n        ### 1. Authentication\n        Register a new user with organization:\n\
    \        ```bash\n        curl -X POST \"http://localhost:8000/api/v1/auth/register\"\
    \ \\\n          -H \"Content-Type: application/json\" \\\n          -d '{\"email\"\
    : \"user@company.com\", \"full_name\": \"John Doe\", \"password\": \"SecurePass123\"\
    , \"organization_name\": \"Acme Corporation\"}'\n        ```\n        \n     \
    \   ### 2. Add JIRA Integration\n        ```bash\n        curl -X POST \"http://localhost:8000/api/v1/integrations\"\
...
    \ and meeting management\n        "
...
openapi: 3.1.0
paths:
...
  /api/v1/auth/check-email:
    post:
      description: Check if email is available for registration
      operationId: check_email_availability_api_v1_auth_check_email_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailCheckRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailCheckResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Check Email Availability
      tags:
      - Authentication
  /api/v1/auth/login:
    post:
      description: Authenticate user and return JWT tokens
      operationId: login_user_api_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Login User
      tags:
      - Authentication
  /api/v1/auth/logout:
    post:
      description: Logout user (client-side token invalidation)
      operationId: logout_user_api_v1_auth_logout_post
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      security:
      - HTTPBearer: []
      summary: Logout User
      tags:
      - Authentication
  /api/v1/auth/me:
    get:
      description: Get current user's profile information
      operationId: get_current_user_profile_api_v1_auth_me_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMeResponse'
          description: Successful Response
      security:
      - HTTPBearer: []
      summary: Get Current User Profile
      tags:
      - Authentication
    put:
      description: Update current user's profile information
      operationId: update_user_profile_api_v1_auth_me_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Update User Profile
      tags:
      - Authentication
  /api/v1/auth/refresh:
    post:
      description: Refresh JWT access token using refresh token
      operationId: refresh_token_api_v1_auth_refresh_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Refresh Token
      tags:
      - Authentication
  /api/v1/auth/register:
    post:
      description: Register a new user account
      operationId: register_user_api_v1_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Register User
      tags:
      - Authentication
  /api/v1/auth/verify:
    get:
      description: Verify if current token is valid
      operationId: verify_token_api_v1_auth_verify_get
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      security:
      - HTTPBearer: []
      summary: Verify Token
      tags:
      - Authentication
...